更新

+++ AdminApp/views.py
@@ -488,12 +488,15 @@ def check_number(number):
             '０': '0', '１': '1', '２': '2', '３': '3', '４': '4',
             '５': '5', '６': '6', '７': '7', '８': '8', '９': '9'
         }))
+        float(number)
 
         numbers = re.findall(r'\d+', number)
+        print(numbers)
 
-
+        
         return True, str(''.join(numbers))
     except ValueError:
+        print("ValueError")
         return False,number
 
 prefix = ""
@@ -519,9 +522,7 @@ def add_kadai_and_answer(request, kadai_id=None):
     else:
         kadai = None
         answers = []
-        
-        
-    
+
     # 問題フォーム
     if request.method == "POST":
         kadai_form = KadaiForm(request.POST, instance=kadai)
@@ -531,6 +532,7 @@ def add_kadai_and_answer(request, kadai_id=None):
             kadai_instance = kadai_form.save(commit=False)  # 問題データを保存前のインスタンス取得
             result,number = check_number(kadai_instance.number)
             if result:
+                print("go")
                 kadai_instance.number = number
                 create_number(kadai_instance)
                 global prefix
@@ -554,6 +556,7 @@ def add_kadai_and_answer(request, kadai_id=None):
 
                     return redirect('AdminApp:admin_kadai_list')  # 遷移先に適宜変更
             else:
+                print("out")
                 kadai_form.add_error('number', '問題番号に数字以外が含まれています。')
 
     else:
@@ -578,20 +581,26 @@ def edit_kadai_and_answers(request, kadai_id):
 
         if kadai_form.is_valid() and answer_formset.is_valid():
             kadai_instance = kadai_form.save(commit=False)
-            create_number(kadai_instance)  # プレフィックス再設定
-            kadai_instance.save()
+            result,number = check_number(kadai_instance.number)
+            if result:
+                kadai_instance.number = number
+                create_number(kadai_instance)  # プレフィックス再設定
+                kadai_instance.save()
 
-            # 解答フォームセットの保存
-            instances = answer_formset.save(commit=False)
-            for instance in instances:
-                instance.kadai = kadai_instance
-                instance.save()
+                # 解答フォームセットの保存
+                instances = answer_formset.save(commit=False)
+                for instance in instances:
+                    instance.kadai = kadai_instance
+                    instance.save()
 
-            # 削除された解答を削除
-            for deleted_instance in answer_formset.deleted_objects:
-                deleted_instance.delete()
+                # 削除された解答を削除
+                for deleted_instance in answer_formset.deleted_objects:
+                    deleted_instance.delete()
 
-            return redirect('AdminApp:admin_kadai_list')
+                return redirect('AdminApp:admin_kadai_list')
+            
+            else:
+                kadai_form.add_error('number', '問題番号に数字以外が含まれています。')
 
     else:
         kadai_form = KadaiForm(instance=kadai)
